LEA (VIPER): Refactored the code structure of the configurator tabs.
