+++ opendut-lea/src/viper_sources/configurator/components/mod.rs
@@ -1,7 +1,3 @@
-mod viper_source_name_input;
-mod viper_source_url_input;
 mod controls;
 
-pub use viper_source_name_input::ViperSourceNameInput;
-pub use viper_source_url_input::ViperSourceUrlInput;
 pub use controls::Controls;
