+++ opendut-lea/src/viper_tests/configurator/tabs/general/mod.rs
@@ -1,8 +1,9 @@
 use leptos::prelude::*;
 use opendut_lea_components::ReadOnlyInput;
-use crate::viper_tests::configurator::components::ViperTestNameInput;
+use crate::viper_tests::configurator::tabs::general::name_input::ViperTestNameInput;
 use crate::viper_tests::configurator::types::UserViperTestRunDescriptor;
 
+pub mod name_input;
 
 #[component]
 pub fn GeneralTab(viper_test_run_descriptor: RwSignal<UserViperTestRunDescriptor>) -> impl IntoView {
