+++ opendut-lea/src/viper_tests/configurator/tabs/source/mod.rs
@@ -1,6 +1,8 @@
 use leptos::prelude::*;
-use crate::viper_tests::configurator::components::ViperTestSourceSelector;
+use crate::viper_tests::configurator::tabs::source::source_selector::ViperTestSourceSelector;
 use crate::viper_tests::configurator::types::UserViperTestRunDescriptor;
+
+pub mod source_selector;
 
 #[component]
 pub fn SourceTab(viper_test_run_descriptor: RwSignal<UserViperTestRunDescriptor>) -> impl IntoView {
